==> 02_linkedin/02_profile_links_clean.ipynb <==
import pandas as pd

df = pd.read_csv('profile_links.csv', names=['link'])

df.head()

df.dropna(inplace=True)
df.drop_duplicates(inplace=True)

len(df)

df.to_csv('profile_links_clean.csv', index=False)
