==> 02_linkedin/05_companies.ipynb <==
from bs4 import BeautifulSoup
import re
import os
import pandas as pd

def read_html_file(file_path):
    with open(file_path, 'r', encoding='utf-8') as file:
        html_text = file.read()
    return html_text


def extract_universal_name(html_text):
    match = re.search(r"universalName:(\d+)", html_text)
    return match.group(1) if match else None


def extract_company_details(html_text):
    soup = BeautifulSoup(html_text, 'html.parser')

    # Extracting company name
    name_tag = soup.find('h1', class_='org-top-card-summary__title')
    name = name_tag.get_text(strip=True) if name_tag else None
    
    company_info = soup.find_all('div', class_='org-top-card-summary-info-list__info-item')
    
    industry = company_info[0].get_text(strip=True) if len(company_info) > 0 else None
    
    location, followers = None, None
    
    if len(company_info) > 1:
        for item in company_info[1:]:
            item = item.get_text(strip=True)
            if 'followers' in item:
                followers = item
            else:
                location = item

    # Extracting employee range
    employees_tag = soup.find('span', class_='t-normal t-black--light link-without-visited-state link-without-hover-state')
    employees = employees_tag.get_text(strip=True) if employees_tag and 'employees' in employees_tag.get_text() else None

    return {
        'name': name,
        'industry': industry,
        'location': location,
        'followers': followers,
        'employees': employees
    }


path = r'D:\github\garage\02_linkedin\download\\'

data = []

for file in os.listdir(path):
    if file.endswith('.html'):
        print(file)
        file_path = path + file
        
        html_text = read_html_file(file_path)
        
        company_id = extract_universal_name(html_text)
        company_details = extract_company_details(html_text)
        company_details['company_id'] = company_id
        
        if company_details['name'] is not None and company_id is not None:
            data.append(company_details)
        

df = pd.DataFrame(data)
df.head()

df[df['company_id'].duplicated()]

df[df['name'].isnull()]

df[df['company_id'].isnull()]

df[df['industry'].isnull()]

df['company_id'] = df['company_id'].astype(int)

df.to_excel('companies.xlsx', index=False)
