==> 02_linkedin/06_companies_list.ipynb <==
import pandas as pd

profiles_df = pd.read_excel('profiles_clean.xlsx')
profiles_df.head()

compainies_df = pd.read_excel('companies.xlsx')
compainies_df.head()

df = compainies_df.join(profiles_df.set_index('company_id'), on='company_id', how='left')
df.head()

df['russians'] = (df['citizenship'] == 'Russia').fillna(0).astype(int)

df.head(30)

compainies_df.columns

columns = ['company_id', 'name', 'industry', 'location', 'followers', 'employees']
df = df.groupby(columns).agg({'headline': 'count', 'russians': 'sum'}).sort_values('russians', ascending=False).reset_index()

df.head()

df.columns = ['company_id', 'name', 'industry', 'location', 'followers', 'employees', 'data_analysts', 'russian_data_analysts']

df['russian_data_analysts_ratio'] = round(100 * df['russian_data_analysts'] / df['data_analysts'], 1)

df.head()

df.to_excel('companies_list.xlsx', index=False)
